[BUGFIX] - Add TF session, handle dif. TF v., fix numerical only
This commit has several fixes. First it now can handle different TensorFlow versions (1 and 2) as some settings are specific for tensorflow 2 or 1. It also adds a session variable because some CF generation frameworks need the TF session. It also adds the DF for the factual points. Finally it fixes a bug when handling numerical only data.