[BUGFIX] Create temp files using time rather than random
Before this commit, a temporary file needed to generate the counterfactual was created using a random function from Numpy. The problem is that the seeds are fixed to allow reproducibility, therefore, the same file name was being created. Now a hash from a timestamp in nanoseconds is create to give a name to a file, this avoids creating in different multiple runs, the same file name and, consequently giving an error.